requirements.txt
# Core dependencies
apache-airflow==2.6.3
apache-airflow-providers-amazon==8.26.0
apache-airflow-providers-docker==3.7.1
apache-airflow-providers-common-sql==1.5.2
psycopg2-binary==2.9.6

# AWS dependencies
boto3==1.28.38
botocore==1.31.38
awswrangler==3.4.1
aws-glue-sessions==1.0.0

# Data processing
dbt-core==1.5.0
dbt-redshift==1.5.0
pandas==2.0.3
pyarrow==12.0.1
fastparquet==2023.7.0

# Visualization
superset==2.1.0

# Terraform & Infrastructure
terraform-parser==0.1.1

# Development
jupyter==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
pre-commit==3.3.3 